+++ stac_pydantic/api/search.py
@@ -1,5 +1,5 @@
 from datetime import datetime as dt
-from typing import Any, Dict, List, Optional, Tuple, Union, cast
+from typing import Any, Dict, List, Optional, Union
 
 from geojson_pydantic.geometries import (
     GeometryCollection,
@@ -10,13 +10,18 @@ from geojson_pydantic.geometries import (
     Point,
     Polygon,
 )
-from pydantic import AfterValidator, BaseModel, Field, TypeAdapter, model_validator
+from pydantic import AfterValidator, BaseModel, Field, model_validator
 from typing_extensions import Annotated
 
 from stac_pydantic.api.extensions.fields import FieldsExtension
 from stac_pydantic.api.extensions.query import Operator
 from stac_pydantic.api.extensions.sort import SortExtension
-from stac_pydantic.shared import BBox, UtcDatetime
+from stac_pydantic.shared import (
+    BBox,
+    str_to_datetimes,
+    validate_bbox,
+    validate_datetime,
+)
 
 Intersection = Union[
     Point,
@@ -27,78 +32,6 @@ Intersection = Union[
     MultiPolygon,
     GeometryCollection,
 ]
-
-SearchDatetime = TypeAdapter(Optional[UtcDatetime])
-
-
-def validate_bbox(v: Optional[BBox]) -> Optional[BBox]:
-    """Validate BBOX value."""
-    if v:
-        # Validate order
-        if len(v) == 4:
-            xmin, ymin, xmax, ymax = cast(Tuple[int, int, int, int], v)
-
-        elif len(v) == 6:
-            xmin, ymin, min_elev, xmax, ymax, max_elev = cast(
-                Tuple[int, int, int, int, int, int], v
-            )
-            if max_elev < min_elev:
-                raise ValueError(
-                    "Maximum elevation must greater than minimum elevation"
-                )
-        else:
-            raise ValueError("Bounding box must have 4 or 6 coordinates")
-
-        # Validate against WGS84
-        if xmin < -180 or ymin < -90 or xmax > 180 or ymax > 90:
-            raise ValueError("Bounding box must be within (-180, -90, 180, 90)")
-
-        if ymax < ymin:
-            raise ValueError("Maximum latitude must be greater than minimum latitude")
-
-    return v
-
-
-def str_to_datetimes(value: str) -> List[Optional[dt]]:
-    # Split on "/" and replace no value or ".." with None
-    values = [v if v and v != ".." else None for v in value.split("/")]
-
-    # Cast because pylance gets confused by the type adapter and annotated type
-    dates = cast(
-        List[Optional[dt]],
-        [
-            # Use the type adapter to validate the datetime strings, strict is necessary
-            # due to pydantic issues #8736 and #8762
-            SearchDatetime.validate_strings(v, strict=True) if v else None
-            for v in values
-        ],
-    )
-    return dates
-
-
-def validate_datetime(v: Optional[str]) -> Optional[str]:
-    """Validate Datetime value."""
-    if v is not None:
-        dates = str_to_datetimes(v)
-
-        # If there are more than 2 dates, it's invalid
-        if len(dates) > 2:
-            raise ValueError(
-                "Invalid datetime range. Too many values. Must match format: {begin_date}/{end_date}"
-            )
-
-        # If there is only one date, duplicate to use for both start and end dates
-        if len(dates) == 1:
-            dates = [dates[0], dates[0]]
-
-        # If there is a start and end date, check that the start date is before the end date
-        if dates[0] and dates[1] and dates[0] > dates[1]:
-            raise ValueError(
-                "Invalid datetime range. Begin date after end date. "
-                "Must match format: {begin_date}/{end_date}"
-            )
-
-    return v
 
 
 class Search(BaseModel):