+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 
 ## Unreleased
 
+- Add datetime validation for collection's time intervals (Must follow [`RFC 3339, section 5.6.`](https://datatracker.ietf.org/doc/html/rfc3339#section-5.6))
+
 ## 3.2.0 (2025-03-20)
 
 - Move `validate_bbox` and `validate_datetime` field validation functions outside the Search class (to enable re-utilization)